Continue the migration
