+++ src/server/trpc/router/_app.ts
@@ -1,18 +1,16 @@
 import { router } from '../trpc'
-import { authRouter } from './auth'
+import { authRouter } from './users/auth'
 import { memberRouter } from './member'
 import { profileRouter } from './profile'
 import { userRouter } from './admin/users'
-import { eventRouter } from './events'
 import { attendanceRouter } from './attendance'
 
 export const appRouter = router({
+  attendance: attendanceRouter,
   auth: authRouter,
-  event: eventRouter,
   member: memberRouter,
   profile: profileRouter,
   user: userRouter,
-  attendance: attendanceRouter,
 })
 
 export type AppRouter = typeof appRouter
