+++ src/server/trpc/router/attendance/index.ts
@@ -4,11 +4,13 @@ import { getAllAttendanceButSelf } from './getAllUsersButSelf'
 import { createEvent } from './createEvent'
 import { getEventInfo } from './getEventInfo'
 import { markAttendance } from './markAttendance'
+import { getEventForUser } from './getEventForUser'
 
 export const attendanceRouter = router({
   createEvent,
   getAllAttendance,
   getAllAttendanceButSelf,
   getEventInfo,
+  getEventForUser,
   markAttendance,
 })
