--- src/server/trpc/router/auth.ts
@@ -1,6 +1,0 @@
-import { router } from '../trpc'
-import { resetPassword } from './auth/post'
-
-export const authRouter = router({
-  resetPassword,
-})
