+++ src/server/trpc/router/users/auth.ts
@@ -1,0 +1,6 @@
+import { router } from '../../trpc'
+import { resetPassword } from './post'
+
+export const authRouter = router({
+  resetPassword,
+})
